feat: :fire: rename totalValue on WeeklyRecap
