+++ app/src/components/WeeklyRecap.vue
